+++ src/ClientInterface.php
@@ -14,6 +14,20 @@ declare(strict_types=1);
 
 namespace Carpediem\Mattermost\Webhook;
 
-class Exception extends \Exception
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\UriInterface;
+
+interface ClientInterface
 {
+    /**
+     * Send a message to a Mattermost Webhook
+     *
+     * @param string|UriInterface $url
+     * @param MessageInterface    $message
+     *
+     * @throws Exception
+     *
+     * @return ResponseInterface
+     */
+    public function notify($url, MessageInterface $message): ResponseInterface;
 }